+++ projects/rocprofiler-sdk/tests/bin/CMakeLists.txt
@@ -45,3 +45,4 @@ add_subdirectory(attachment-test)
 add_subdirectory(hip-host)
 add_subdirectory(module-loading-test)
 add_subdirectory(late-start-tracing)
+add_subdirectory(thread-trace)
