+++ projects/rocprofiler-sdk/tests/bin/thread-trace/kernel_lds.cpp
