+++ projects/rocprofiler-sdk/tests/bin/thread-trace/main.cpp
