+++ projects/rocprofiler-sdk/tests/tools/CMakeLists.txt
@@ -54,6 +54,51 @@ add_custom_command(
     WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
     COMMENT "Stripping rocprofiler-sdk-c-tool...")
 
+# thread-trace tool libraries
+set(THREAD_TRACE_TOOL_LINK_LIBRARIES
+    rocprofiler-sdk::rocprofiler-sdk rocprofiler-sdk::tests-build-flags
+    rocprofiler-sdk::tests-common-library)
+
+# single dispatch thread trace tool
+add_library(thread-trace-single-tool SHARED)
+target_sources(thread-trace-single-tool
+               PRIVATE thread-trace-single-tool.cpp thread-trace-callbacks.cpp)
+target_link_libraries(thread-trace-single-tool PRIVATE ${THREAD_TRACE_TOOL_LINK_LIBRARIES})
+set_target_properties(
+    thread-trace-single-tool
+    PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/rocprofiler-sdk"
+               INSTALL_RPATH "\$ORIGIN:\$ORIGIN/..")
+
+# multi dispatch thread trace tool
+add_library(thread-trace-multi-tool SHARED)
+target_sources(thread-trace-multi-tool
+               PRIVATE thread-trace-multi-tool.cpp thread-trace-callbacks.cpp)
+target_link_libraries(thread-trace-multi-tool PRIVATE ${THREAD_TRACE_TOOL_LINK_LIBRARIES})
+set_target_properties(
+    thread-trace-multi-tool
+    PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/rocprofiler-sdk"
+               INSTALL_RPATH "\$ORIGIN:\$ORIGIN/..")
+
+# agent (device) thread trace tool
+add_library(thread-trace-agent-tool SHARED)
+target_sources(thread-trace-agent-tool
+               PRIVATE thread-trace-agent-tool.cpp thread-trace-callbacks.cpp)
+target_link_libraries(thread-trace-agent-tool PRIVATE ${THREAD_TRACE_TOOL_LINK_LIBRARIES})
+set_target_properties(
+    thread-trace-agent-tool
+    PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/rocprofiler-sdk"
+               INSTALL_RPATH "\$ORIGIN:\$ORIGIN/..")
+
+# triple buffer thread trace tool
+add_library(thread-trace-triple-buffer-tool SHARED)
+target_sources(thread-trace-triple-buffer-tool PRIVATE thread-trace-triple-buffer-tool.cpp)
+target_link_libraries(thread-trace-triple-buffer-tool
+                      PRIVATE ${THREAD_TRACE_TOOL_LINK_LIBRARIES})
+set_target_properties(
+    thread-trace-triple-buffer-tool
+    PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/rocprofiler-sdk"
+               INSTALL_RPATH "\$ORIGIN:\$ORIGIN/..")
+
 # tool library supporting late initialization in the application
 add_library(rocprofiler-sdk-late-start-tool SHARED)
 target_sources(rocprofiler-sdk-late-start-tool PRIVATE late-start-tool.cpp)
