+++ projects/rocprofiler-sdk/tests/tools/thread-trace-callbacks.hpp
