+++ projects/rocprofiler-sdk/tests/tools/thread-trace-single-tool.cpp
@@ -25,7 +25,7 @@
 #    undef NDEBUG
 #endif
 
-#include "trace_callbacks.hpp"
+#include "thread-trace-callbacks.hpp"
 
 namespace ATTTest
 {
