+++ projects/rocprofiler-sdk/tests/tools/thread-trace-triple-buffer-tool.cpp
